Updated the scid SlackBuild to 4.7.0
The update includes a number of changes which implement the features
found in the pre-built packages. The updates include the download of
the photo files, ratings/spelling file, sounds, and scidlet opening
book.

We've added the ENGINES build option which will enable the
installation of the phalanx-scid, scidlet, and togaII1.2.1a engines.

There are several patches which update the Tcl source to set the
default locations and settings for the scid.eco file, spellcheck file,
and engine directories. These features are now automatically found
when starting up a fresh scid instance.