network/monkey: Don't package /var/run/monkey
Since /var/run is a link to /run (and /run is tmpfs), there's
no point packaging any subdir of /run or /var/run.

It's highly likely that further changes will be needed for
this build. Among those are:
1) create /var/run/monkey/ during init (rc.monkey)
2) change ownership of /var/run/monkey/ to apache:apache during init

Unrelated to the above, but also needed if this app uses
logrotate to rotate its logs: the logrotate file needs to
specify the user and group to switch to for the rotation process.
See e.g. the logrotate file in network/yadifa/

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>