Fixed issue missing url in runtime update.
There are now empty url in the update schema which breaks the runtime
download in lutris. This is not really a full fix, since it skips these urls.

Currently, it is only the icons that are skipped. To fix completely,
the downloader should instead download a separate JSON file which
contains the urls to the icons. This support was added in newer
version of lutris.